rag/requirements.txt
flask
pymongo
python-docx
langchain
chromadb
sentence-transformers
numpy
requests
gunicorn